Update README, add warning for hpjob missing files

+++ jobs/haskellplatform/hpjob.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 cd "`dirname "$0"`"
-set -e -x
+set -e
 HP_VERSION=2012.4.0.0
 
 mkdir -p job/bin
@@ -18,10 +18,20 @@ if [ ! -f "job/${GHC_OPT_ARCHIVE}" ]; then
     exit 1
 fi
 
+mkdir -p job/debs
+if [ ! -f job/debs/x11-common* ]; then
+    echo "This job requires debian archives to install"
+    echo "the haskell platform from the source tar."
+    echo "You can download the required binary packages by"
+    echo "running the debs job on a network enabled jobrunner"
+    echo "You should then place the haskellpatform .deb files in job/debs"
+    exit 1
+fi
+
 sed -s 's/HP_VERSION/'${HP_VERSION}'/g' < construir.template > job/bin/construir
 sed -i 's/GHC_OPT_ARCHIVE/'${GHC_OPT_ARCHIVE}'/g' job/bin/construir
 chmod a+x job/bin/construir
 
 mkdir -p job/output
-../mkjob.py --extra-space 1000 job
+../mkjob.py --extra-space 200 job
